requirements/base.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.3
fastapi-users-db-sqlalchemy==6.0.1
bcrypt==4.0.1
pycountry==23.12.11
SQLAlchemy-Utils==0.41.1
fastapi-cache2==0.2.1
redis==5.0.1
requests==2.31.0